backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
bcrypt>=4.1.3,<6
beanie==1.26.0
motor==3.6.0
python-jose==3.3.0
apscheduler==3.10.4
pydantic[email]==2.12.4
pydantic-settings==2.5.2
pytest==8.3.2
httpx==0.27.2
python-dotenv==1.0.1
openpyxl==3.1.5
reportlab==4.2.5
